Added exception

+++ scrandom.py
@@ -63,6 +63,8 @@ def clear_old_files(force=False):
     ensure_dir_exists()
     with os.scandir(DIRECTORY) as it:
         for entry in it:
+            if entry.name == "README.md":
+                continue
             if force:
                 print(f"Force-removing {entry.name}")
                 os.remove(entry)
